+++ firestore-stripe-subscriptions/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 0.1.13 - 2021-06-17
+
+[fix] Add a `prices` and an `items` array to the one-time payment docs in the `payments` collection. The `prices` array holds Firestore references for the prices that make up this payment, and the `items` array includes the full line items of the Stripe Checkout session.
+
 ## Version 0.1.12 - 2021-04-29
 
 [feat] Sync customer email address to Cloud Firestore. (#157)
@@ -264,7 +268,7 @@ type: "one_time" | "recurring";
 
 You can create customer-facing promotion codes in the [Stripe Dashboard](https://dashboard.stripe.com/coupons/create). Refer to the [docs](https://stripe.com/docs/billing/subscriptions/discounts/codes) for a detailed guide on how to set these up.
 
-To show the promotion code redemption box on the checkout page, set `allow_promotion_codes: true` when creating the `checkout_session` document:
+To show the promotion code redemption box on the checkout page, set `allow_promotion_codes: true` when creating the `checkout_sessions` document:
 
 ```js
 const docRef = await db